+++ projects/command/src/command_pattern_implementation/concerns/inputs_type.rb
@@ -12,6 +12,10 @@ module Foobara
               unregister_possible_error_if_registered(possible_error)
             end
 
+            if defined?(@inputs_association_paths)
+              remove_instance_variable(:@inputs_association_paths)
+            end
+
             type = type_for_declaration(...)
 
             if type.extends?(BuiltinTypes[:model]) && !type.extends?(BuiltinTypes[:entity])