+++ projects/command/src/command_pattern_implementation/concerns/runtime.rb
@@ -7,6 +7,13 @@ module Foobara
         class CannotHaltWithoutAddingErrors < StandardError; end
         class Halt < StandardError; end
 
+        class NotFoundError < Foobara::DataError
+          context do
+            entity_class :string, :required
+            criteria :duck
+          end
+        end
+
         module ClassMethods
           def run(...)
             new(...).run
@@ -81,7 +88,41 @@ module Foobara
         end
 
         def validate_records
-          # noop
+          self.class.inputs_association_paths&.each do |data_path|
+            if data_path.last == :"#"
+              records = data_path.values_at(@inputs)
+
+              records.each.with_index do |record, index|
+                if record&.persisted? && !record.loaded?
+                  begin
+                    record.class.load(record)
+                  rescue Foobara::Entity::NotFoundError => e
+                    add_input_error(
+                      [*data_path.path[..-2], index],
+                      CommandPatternImplementation::NotFoundError,
+                      criteria: e.criteria,
+                      entity_class: record.class.model_type.scoped_full_name
+                    )
+                  end
+                end
+              end
+            else
+              record = data_path.value_at(@inputs)
+
+              if record&.persisted? && !record.loaded?
+                begin
+                  record.class.load(record)
+                rescue Foobara::Entity::NotFoundError => e
+                  add_input_error(
+                    data_path.to_s,
+                    CommandPatternImplementation::NotFoundError,
+                    criteria: e.criteria,
+                    entity_class: record.class.model_type.scoped_full_name
+                  )
+                end
+              end
+            end
+          end
         end
 
         def validate